moved utility functions to util directory for reusability
