+++ util/format.js
@@ -1,0 +1,10 @@
+// format input value
+function format(value) {
+  // separating parameter assignment and parseFloat operation for consistent outcome
+  value = parseFloat(value) // outputs: number
+  value = value.toFixed(2) // outputs: string
+  value = parseFloat(value) // outputs: number
+  return value // number
+}
+
+module.exports = format;
